// src/app/Models/model-entities.ts
export class UserDataEntities {
    UserId: number;
    FirstName: string;
    LastName: string;
    EmpId: number;
    UserStatus: string;
    AddDate: string;
    UpdtDate: string;
}

export class ProjectEntities {
    ProjectId: number;
    ProjectName: string;
    ProjectPriority: number;
    DateReqd: string;
    StartDate: string;
    EndDate: string;
    ManagerId: number;
    ProjectStatus: string;
    AddDate: string;
    UpdtDate: string;
}